feat: timestamp in enrich message and packet

+++ common/packet.go
@@ -1,10 +1,10 @@
 package common
 
 type Packet struct {
-	SrcIp    string
-	SrcPort  string
-	DestIp   string
-	DestPort string
-	Size     uint16
+	SrcIp     string
+	SrcPort   string
+	DestIp    string
+	DestPort  string
+	Size      uint32
 	TimeStamp string
 }
